Make buffer size/count for workers configurable
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/branches/issue-503@13570 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/turbine/worker.c
@@ -39,7 +39,7 @@ static void task_error(Tcl_Interp* interp, int tcl_rc, char* command);
   TODO: priority isn't inherited from parent tasks
  */
 turbine_code
-turbine_worker_loop(Tcl_Interp* interp, void* buffer, size_t buffer_size,
+turbine_worker_loop(Tcl_Interp* interp, void* buffer, int buffer_size,
                     int work_type)
 {
   int rc;
