+++ code/src/turbine/worker.h
@@ -10,7 +10,7 @@
 #define WORKER_H
 
 turbine_code turbine_worker_loop(Tcl_Interp* interp,
-                                 void* buffer, size_t buffer_size,
+                                 void* buffer, int buffer_size,
                                  int work_type);
 
 #endif
